Move workers to individual files
