+++ lib/meetup_bot/workers/post_to_slack_worker.ex
@@ -5,7 +5,6 @@ defmodule MeetupBot.PostToSlackWorker do
 
   alias MeetupBot.MeetupCache
   alias MeetupBot.Slack
-  alias MeetupBot.Meetup
   alias OpenTelemetry.Tracer
 
   def cron do
@@ -28,25 +27,3 @@ defmodule MeetupBot.PostToSlackWorker do
   end
 end
 
-defmodule MeetupBot.MeetupCacheWorker do
-  use Oban.Worker
-
-  require OpenTelemetry.Tracer
-
-  alias MeetupBot.MeetupCache
-  alias MeetupBot.Meetup
-  alias MeetupBot.GDG
-  alias OpenTelemetry.Tracer
-
-  @impl true
-  def perform(%Oban.Job{}) do
-    Tracer.with_span "oban.perform" do
-      Tracer.set_attributes([{:worker, "MeetupCacheWorker"}])
-
-      (Meetup.fetch_upcoming_meetups() ++ GDG.fetch_live_events())
-      |> MeetupCache.update()
-
-      :ok
-    end
-  end
-end
